% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_assumptions.R
\name{anova_assumptions}
\alias{anova_assumptions}
\title{Diagnose ANOVA Assumptions}
\usage{
anova_assumptions(x)
}
\arguments{
\item{x}{a dataframe to check anova assumptions on (for location and frame)}
}
\value{
4 plots that allow us to diagnose ANOVA assumptions
}
\description{
Diagnose ANOVA Assumptions
}
