% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_results.R
\name{aov_results}
\alias{aov_results}
\title{ANOVA Results}
\usage{
aov_results(x)
}
\arguments{
\item{x}{a dataframe to get the ANOVA results of (for location and frame)}
}
\value{
A dataframe that contains the summary statistics of the ANOVA
}
\description{
ANOVA Results
}
