% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean Diabetes Data}
\usage{
clean_data(x)
}
\arguments{
\item{x}{a dataframe to clean data on.}
}
\value{
A dataframe that cleans out na values, drops empty columns, and
selects the variables of interest.
}
\description{
Clean Diabetes Data
}
