% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Conditions Data from ctrialsgov}
\format{
\subsection{\code{designs}}{

A data frame with 814,683 rows and 4 columns:
\describe{
...
}
}
}
\usage{
conditions
}
\description{
Conditions data for BIS620
}
\keyword{datasets}
