% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adverse_events_histogram.R
\name{create_adverse_events_histogram}
\alias{create_adverse_events_histogram}
\title{Plots histogram of adverse events from the trials}
\usage{
create_adverse_events_histogram(d, con)
}
\arguments{
\item{d}{a dataset given to the function for merging}

\item{con}{a duckdkb connection that will help with computational efficiency
Otherwise, it will take a long time to run this function. SQL helps a lot.}
}
\value{
ggplot object that shows a histogram of the distribution of
adverse events
}
\description{
This function creates a histogram of the adverse events in the
trial
}
