% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_conditions_histogram.R
\name{create_conditions_histogram}
\alias{create_conditions_histogram}
\title{Plots histogram of conditions from the trials}
\usage{
create_conditions_histogram(d, conditions)
}
\arguments{
\item{d}{a dataset given to the function for merging}

\item{conditions}{a dataset with the conditions}
}
\value{
ggplot object that shows a histogram of the distribution of
conditions
}
\description{
This function creates a histogram of the conditions in the trial
}
