% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_endpoint_histogram.R
\name{create_endpoint_histogram}
\alias{create_endpoint_histogram}
\title{Plots histogram of endpoints from the trials}
\usage{
create_endpoint_histogram(d, endpoints)
}
\arguments{
\item{d}{a dataset given to the function for merging}

\item{endpoints}{a dataset with the endpoints}
}
\value{
ggplot object that shows a histogram of the distribution of
endpoints
}
\description{
This function creates a histogram of the endpoints in the trial
}
