% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intervention_histogram.R
\name{create_intervention_histogram}
\alias{create_intervention_histogram}
\title{Plots histogram of intervention study types from the trials}
\usage{
create_intervention_histogram(d, designs)
}
\arguments{
\item{d}{the dataframe of the query result}

\item{designs}{the handle referencing the "designs" database table}
}
\value{
ggplot object that shows a histogram of the distribution of
intervention study types
}
\description{
Displays a histogram to show the distribution of interventional
study types from the trials in a query
}
