% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_observational_histogram.R
\name{create_observational_histogram}
\alias{create_observational_histogram}
\title{Plots histogram of observational study types from the trials}
\usage{
create_observational_histogram(d, designs)
}
\arguments{
\item{d}{the dataframe of the query result}

\item{designs}{the handle referencing the "designs" database table}
}
\value{
ggplot object that shows a histogram of the distribution of
observational study types
}
\description{
Displays a histogram to show the distribution of observational
study types from the trials in a query
}
