% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phase_histogram.R
\name{create_phase_histogram}
\alias{create_phase_histogram}
\title{Plots histogram of phases from the trials}
\usage{
create_phase_histogram(d, studies)
}
\arguments{
\item{d}{a dataset given to the function}

\item{studies}{a dataset with the phase information}
}
\value{
ggplot object that shows a histogram of the distribution of
phases
}
\description{
This function creates a histogram of the phases in the trial
}
