% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\docType{data}
\name{designs}
\alias{designs}
\title{Designs Data from ctrialsgov}
\format{
\subsection{\code{designs}}{

A data frame with 465,815 rows and 14 columns:
\describe{
...
}
}
}
\usage{
designs
}
\description{
Designs data for BIS620
}
\keyword{datasets}
