% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Data from Dr. John Schorling from the University of Virginia School
of Medicine}
\format{
\subsection{\code{diabetes}}{

A data frame with 403 rows and 19 columns:
\describe{
\item{glyhb}{glycosolated hemoglobin level}
\item{ratio}{ratio of cholesterol to high-density lipoprotein}
\item{frame}{body frame size (low, medium, high)}
\item{location}{living location (Louisa or Birmingham)}
...
}
}
}
\usage{
diabetes
}
\description{
Diabetes data for BIS 620 Final
}
\keyword{datasets}
