% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\docType{data}
\name{endpoints}
\alias{endpoints}
\title{Endpoints Data from ctrialsgov}
\format{
\subsection{\code{endpoints}}{

A data frame with 16,487 rows and 3 columns:
\describe{
...
}
}
}
\usage{
endpoints
}
\description{
Endpoints data for BIS620
}
\keyword{datasets}
