% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_hist.R
\name{explore_hist}
\alias{explore_hist}
\title{Explore Diabetes Data with Histogram}
\usage{
explore_hist(x)
}
\arguments{
\item{x}{a dataframe to perform exploratory data analysis on}
}
\value{
two histograms of the quantitative variables
}
\description{
Explore Diabetes Data with Histogram
}
