% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_assumptions.R
\name{lr_assumptions}
\alias{lr_assumptions}
\title{Diagnose Linear Regression Assumptions}
\usage{
lr_assumptions(x, log = FALSE)
}
\arguments{
\item{x}{a dataframe to check linear regression assumptions on (for the
ratio of cholesterol to high-density lipoproteins)}
}
\value{
4 plots that allow us to diagnose linear regression assumptions
}
\description{
Diagnose Linear Regression Assumptions
}
