% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_results.R
\name{lr_results}
\alias{lr_results}
\title{Linear Regression Results}
\usage{
lr_results(x)
}
\arguments{
\item{x}{a dataframe to get the linear regression results of (for location
and frame)}
}
\value{
A dataframe that contains the summary statistics of the linear
regression
}
\description{
Linear Regression Results
}
