% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_kwds.R
\name{query_kwds}
\alias{query_kwds}
\title{Query the Keywords from the Given Table}
\usage{
query_kwds(tbl, kwds, column, ignore_case = TRUE, match_all = FALSE)
}
\arguments{
\item{tbl}{a dataset to query on}

\item{kwds}{word(s) that will be used to filter the dataset}

\item{column}{name of the column/variable within the dataset to query the
words from}

\item{ignore_case}{should the cases be ignored. default is 'TRUE'}

\item{match_all}{should the words be a complete match. default is 'FALSE'}
}
\value{
a data frame with the specified query
}
\description{
This function queries a given data set using keywords given to the function.
}
