% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies_subset.R
\docType{data}
\name{studies_subset}
\alias{studies_subset}
\title{A Subset of Data from the Studies Dataset from ctrialsgov}
\format{
\subsection{\code{studies_subset}}{

A data frame with 1,000 rows and 70 columns:
\describe{
...
}
}
}
\usage{
studies_subset
}
\description{
Subsetted data for BIS620
}
\keyword{datasets}
